chore(editor-state): refactoring file paths
