+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockDecorators/buttons/interaction.ts
@@ -1,5 +1,5 @@
 import { EventDispatcher } from '../../../../types';
-import findButtonAtViewportCoordinates from '../../../../helpers/findButtonAtViewportCoordinates';
+import findButtonAtViewportCoordinates from '../../../../helpers/finders/findButtonAtViewportCoordinates';
 
 import type { State, Switch, CodeBlockGraphicData } from '../../../../types';
 import type { DataStructure } from '@8f4e/compiler';
