+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockDecorators/pianoKeyboard/interaction.ts
@@ -2,7 +2,7 @@ import { StateManager } from '@8f4e/state-manager';
 
 import { CodeBlockClickEvent } from '../../codeBlockDragger';
 import { EventDispatcher } from '../../../../types';
-import findPianoKeyAtViewportCoordinates from '../../../../helpers/findPianoKeyboardAtViewportCoordinates';
+import findPianoKeyAtViewportCoordinates from '../../../../helpers/finders/findPianoKeyboardAtViewportCoordinates';
 import { insertCodeAfterLine, replaceCode } from '../../../../helpers/multiLineCodeParser';
 
 import type { State } from '../../../../types';
