+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockDecorators/switches/interaction.ts
@@ -1,5 +1,5 @@
 import { EventDispatcher } from '../../../../types';
-import findSwitchAtViewportCoordinates from '../../../../helpers/findSwitchAtViewportCoordinates';
+import findSwitchAtViewportCoordinates from '../../../../helpers/finders/findSwitchAtViewportCoordinates';
 
 import type { State, CodeBlockGraphicData } from '../../../../types';
