+++ packages/editor/packages/editor-state/src/effects/menu/contextMenu.ts
@@ -3,7 +3,7 @@ import { StateManager } from '@8f4e/state-manager';
 import * as menus from './menus';
 
 import { EventDispatcher } from '../../types';
-import findCodeBlockAtViewportCoordinates from '../../helpers/findCodeBlockAtViewportCoordinates';
+import findCodeBlockAtViewportCoordinates from '../../helpers/finders/findCodeBlockAtViewportCoordinates';
 
 import type { ContextMenuItem, State } from '../../types';
