+++ packages/editor/packages/editor-state/src/helpers/finders/findClosestCodeBlockInDirection.ts
@@ -1,4 +1,4 @@
-import type { CodeBlockGraphicData } from '../types';
+import type { CodeBlockGraphicData } from '../../types';
 
 /**
  * Direction types for navigation
