+++ packages/editor/packages/editor-state/src/helpers/finders/findCodeBlockAtViewportCoordinates.ts
@@ -1,4 +1,4 @@
-import type { CodeBlockGraphicData, GraphicHelper } from '../types';
+import type { CodeBlockGraphicData, GraphicHelper } from '../../types';
 
 export default function findCodeBlockAtViewportCoordinates(
 	graphicHelper: GraphicHelper,
