+++ packages/editor/packages/editor-state/src/helpers/finders/findSwitchAtViewportCoordinates.ts
@@ -1,4 +1,4 @@
-import type { CodeBlockGraphicData, GraphicHelper, Switch } from '../types';
+import type { CodeBlockGraphicData, GraphicHelper, Switch } from '../../types';
 
 export default function findSwitchAtViewportCoordinates(
 	graphicHelper: GraphicHelper,
